requirements.txt
pyyaml==3.12
